requirements.txt
attrs
google-cloud-storage
google-cloud-container
kubernetes
